requirements.txt
python-telegram-bot[job-queue]==20.1
aiohttp
requests==2.32.3
# Supabase HTTP implementation uses aiohttp & requests natively. No admin SDKs needed.
nest_asyncio==1.6.0
flask
keep_alive==0.0.1
python-dotenv
